@@ -6,7 +6,8 @@
     "homepage": "https://github.com/Ashus/google-addressbook",
     "license": "GPL-3.0+",
     "require": {
-        "php": ">=5.6",
+        "php": ">=7.0",
+        "ext-json": "*",
         "roundcube/plugin-installer": "^0.3",
         "google/apiclient": "^2.12"
     },
